Optimise code to split each template into it's own class
Allows for easier code maintenance.
