+++ src/main/kotlin/uk/gov/justice/digital/hmpps/notificationsalertsvsip/interfaces/EmailNotificationHandler.kt
@@ -1,0 +1,8 @@
+package uk.gov.justice.digital.hmpps.notificationsalertsvsip.interfaces
+
+import uk.gov.justice.digital.hmpps.notificationsalertsvsip.dto.SendEmailNotificationDto
+import uk.gov.justice.digital.hmpps.notificationsalertsvsip.dto.visit.scheduler.VisitDto
+
+interface EmailNotificationHandler {
+  fun handle(visit: VisitDto): SendEmailNotificationDto
+}
